Moved keystore to the ehcache plugin
