--- modules/roaringbitmap-keystore/build.gradle
@@ -1,25 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-opensearchplugin {
-  description 'Roaring bitmap-based implementation of a key lookup store, used for tiered caching.'
-  classname 'org.opensearch.cache.RoaringBitmapKeystoreModulePlugin'
-  hasClientJar = true
-}
-
-dependencies {
-  api 'org.roaringbitmap:RoaringBitmap:0.9.49'
-  runtimeOnly 'org.roaringbitmap:shims:0.9.49'
-}
-
-tasks.withType(Javadoc).configureEach {
-  failOnError false
-  options.addStringOption('Xdoclint:none', '-quiet')
-  options.addStringOption('encoding', 'UTF-8')
-  options.addStringOption('charSet', 'UTF-8')
-}
