--- modules/roaringbitmap-keystore/src/main/java/org/opensearch/cache/RoaringBitmapKeystoreModulePlugin.java
@@ -1,13 +1,0 @@
-/*
- * SPDX-License-Identifier: Apache-2.0
- *
- * The OpenSearch Contributors require contributions made to
- * this file be licensed under the Apache-2.0 license or a
- * compatible open source license.
- */
-
-package org.opensearch.cache;
-
-import org.opensearch.plugins.Plugin;
-
-public class RoaringBitmapKeystoreModulePlugin extends Plugin {}
