+++ plugins/cache-ehcache/build.gradle
@@ -18,7 +18,12 @@ opensearchplugin {
 }
 
 dependencies {
+  // for ehcache
   api "org.ehcache:ehcache:${versions.ehcache}"
+
+  // For roaring bitmaps
+  api 'org.roaringbitmap:RoaringBitmap:0.9.49'
+  runtimeOnly 'org.roaringbitmap:shims:0.9.49'
 }
 
 thirdPartyAudit {
