+++ plugins/cache-ehcache/src/main/java/org/opensearch/cache/keystore/KeyStoreStats.java
@@ -6,7 +6,7 @@
  * compatible open source license.
  */
 
-package org.opensearch.cache;
+package org.opensearch.cache.keystore;
 
 import org.opensearch.common.metrics.CounterMetric;
